[wetek] Fix receiving maxint(64bit) from Python in eActionMap, scale to 32bit (reverted from commit c38817eb74453efd2f353430b7263b9d5df1bdcd)

+++ lib/actions/action.h
@@ -21,11 +21,11 @@ public:
 #ifndef SWIG
 	eActionMap();
 	~eActionMap();
-	void bindAction(const std::string &context, long long int priority, int id, eWidget *widget);
+	void bindAction(const std::string &context, int priority, int id, eWidget *widget);
 	void unbindAction(eWidget *widget, int id);
 #endif
 
-	void bindAction(const std::string &context, long long int priority, SWIG_PYOBJECT(ePyObject) function);
+	void bindAction(const std::string &context, int priority, SWIG_PYOBJECT(ePyObject) function);
 	void unbindAction(const std::string &context, SWIG_PYOBJECT(ePyObject) function);
 
 	void bindKey(const std::string &domain, const std::string &device, int key, int flags, const std::string &context, const std::string &action);
